requirements.txt
aiodocker==0.21.0
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
CacheControl==0.12.11
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cleo==2.0.1
click==8.1.3
crashtest==0.4.1
cryptography==39.0.0
distlib==0.3.6
dnspython==2.2.1
dulwich==0.20.50
email-validator==1.3.0
exceptiongroup==1.1.0
fastapi==0.89.0
filelock==3.9.0
frozenlist==1.3.3
h11==0.14.0
html5lib==1.1
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
importlib-metadata==4.13.0
importlib-resources==5.10.2
iniconfig==2.0.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
jeepney==0.8.0
Jinja2==3.1.2
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
MarkupSafe==2.1.1
more-itertools==9.0.0
motor==3.1.1
msgpack==1.0.4
multidict==6.0.4
mypy-extensions==0.4.3
odmantic==0.9.2
orjson==3.8.4
packaging==22.0
pathspec==0.10.3
pexpect==4.8.0
pkg_resources==0.0.0
pkginfo==1.9.6
pkgutil_resolve_name==1.3.10
platformdirs==2.6.2
pluggy==1.0.0
poetry==1.3.1
poetry-core==1.4.0
poetry-plugin-export==1.2.0
ptyprocess==0.7.0
pycparser==2.21
pydantic==1.10.4
pymongo==4.3.3
pyrsistent==0.19.3
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
rapidfuzz==2.13.7
requests==2.28.1
requests-toolbelt==0.10.1
rfc3986==1.5.0
SecretStorage==3.3.3
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
tomli==2.0.1
tomlkit==0.11.6
trove-classifiers==2022.12.22
typing_extensions==4.4.0
ujson==5.7.0
urllib3==1.26.13
uvicorn==0.20.0
uvloop==0.17.0
virtualenv==20.17.1
watchfiles==0.18.1
webencodings==0.5.1
websockets==10.4
yarl==1.8.2
zipp==3.11.0
